[HUDI-4584] Cleaning up Spark utilities (#6351)
Cleans up Spark utilities and removes duplication
