+++ hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/keygen/BuiltinKeyGenerator.java
@@ -451,7 +451,7 @@ public abstract class BuiltinKeyGenerator extends BaseKeyGenerator implements Sp
     private final SparkRowSerDe rowSerDe;
 
     SparkRowConverter(StructType schema) {
-      this.rowSerDe = HoodieSparkUtils.getDeserializer(schema);
+      this.rowSerDe = HoodieSparkUtils.getCatalystRowSerDe(schema);
       this.avroConverter = AvroConversionUtils.createConverterToAvro(schema, STRUCT_NAME, NAMESPACE);
     }
