+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/TestConvertFilterToCatalystExpression.scala
@@ -17,7 +17,8 @@
 
 package org.apache.hudi
 
-import org.apache.hudi.HoodieSparkUtils.convertToCatalystExpression
+import org.apache.spark.sql.HoodieCatalystExpressionUtils
+import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types._
 import org.junit.jupiter.api.Assertions.assertEquals
@@ -34,6 +35,16 @@ class TestConvertFilterToCatalystExpression {
     fields.append(StructField("price", DoubleType, nullable = true))
     fields.append(StructField("ts", IntegerType, nullable = false))
     StructType(fields)
+  }
+
+  private def convertToCatalystExpression(filters: Array[Filter],
+                                  tableSchema: StructType): Option[Expression] = {
+    val expressions = filters.map(HoodieCatalystExpressionUtils.convertToCatalystExpression(_, tableSchema))
+    if (expressions.nonEmpty && expressions.forall(p => p.isDefined)) {
+      Some(expressions.map(_.get).reduce(org.apache.spark.sql.catalyst.expressions.And))
+    } else {
+      None
+    }
   }
 
   @Test
@@ -74,7 +85,7 @@ class TestConvertFilterToCatalystExpression {
     } else {
       expectExpression
     }
-    val exp = convertToCatalystExpression(filter, tableSchema)
+    val exp = HoodieCatalystExpressionUtils.convertToCatalystExpression(filter, tableSchema)
     if (removeQuotesIfNeed == null) {
       assertEquals(exp.isEmpty, true)
     } else {
