+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/TestHoodieSparkUtils.scala
@@ -19,19 +19,14 @@
 package org.apache.hudi
 
 import org.apache.avro.generic.GenericRecord
-import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.Path
 import org.apache.hudi.testutils.DataSourceTestUtils
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{Row, SparkSession}
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.io.TempDir
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 
-import java.io.File
-import java.nio.file.Paths
 import scala.collection.JavaConverters
 
 class TestHoodieSparkUtils {
@@ -88,58 +83,6 @@ class TestHoodieSparkUtils {
         assertFalse(vsMock.isSpark3_0)
         assertFalse(vsMock.isSpark3_1)
     }
-  }
-
-  @Test
-  def testGlobPaths(@TempDir tempDir: File): Unit = {
-    val folders: Seq[Path] = Seq(
-      new Path(Paths.get(tempDir.getAbsolutePath, "folder1").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, "folder2").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, ".hoodie").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, ".hoodie", "metadata").toUri)
-    )
-
-    val files: Seq[Path] = Seq(
-      new Path(Paths.get(tempDir.getAbsolutePath, "folder1", "file1").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, "folder1", "file2").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, "folder2", "file3").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, "folder2","file4").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, ".hoodie","metadata", "file5").toUri),
-      new Path(Paths.get(tempDir.getAbsolutePath, ".hoodie","metadata", "file6").toUri)
-    )
-
-    folders.foreach(folder => new File(folder.toUri).mkdir())
-    files.foreach(file => new File(file.toUri).createNewFile())
-
-    var paths = Seq(tempDir.getAbsolutePath + "/*")
-    var globbedPaths = HoodieSparkUtils.checkAndGlobPathIfNecessary(paths,
-      new Path(paths.head).getFileSystem(new Configuration()))
-    assertEquals(folders.filterNot(entry => entry.toString.contains(".hoodie"))
-      .sortWith(_.toString < _.toString), globbedPaths.sortWith(_.toString < _.toString))
-
-    paths = Seq(tempDir.getAbsolutePath + "/*/*")
-    globbedPaths = HoodieSparkUtils.checkAndGlobPathIfNecessary(paths,
-      new Path(paths.head).getFileSystem(new Configuration()))
-    assertEquals(files.filterNot(entry => entry.toString.contains(".hoodie"))
-      .sortWith(_.toString < _.toString), globbedPaths.sortWith(_.toString < _.toString))
-
-    paths = Seq(tempDir.getAbsolutePath + "/folder1/*")
-    globbedPaths = HoodieSparkUtils.checkAndGlobPathIfNecessary(paths,
-      new Path(paths.head).getFileSystem(new Configuration()))
-    assertEquals(Seq(files(0), files(1)).sortWith(_.toString < _.toString),
-      globbedPaths.sortWith(_.toString < _.toString))
-
-    paths = Seq(tempDir.getAbsolutePath + "/folder2/*")
-    globbedPaths = HoodieSparkUtils.checkAndGlobPathIfNecessary(paths,
-      new Path(paths.head).getFileSystem(new Configuration()))
-    assertEquals(Seq(files(2), files(3)).sortWith(_.toString < _.toString),
-      globbedPaths.sortWith(_.toString < _.toString))
-
-    paths = Seq(tempDir.getAbsolutePath + "/folder1/*", tempDir.getAbsolutePath + "/folder2/*")
-    globbedPaths = HoodieSparkUtils.checkAndGlobPathIfNecessary(paths,
-      new Path(paths.head).getFileSystem(new Configuration()))
-    assertEquals(files.filterNot(entry => entry.toString.contains(".hoodie"))
-      .sortWith(_.toString < _.toString), globbedPaths.sortWith(_.toString < _.toString))
   }
 
   @Test
