+++ hudi-spark-datasource/hudi-spark3-common/src/main/scala/org/apache/spark/sql/HoodieSpark3CatalystPlanUtils.scala
@@ -85,8 +85,4 @@ trait HoodieSpark3CatalystPlanUtils extends HoodieCatalystPlansUtils {
                                 query: LogicalPlan, overwrite: Boolean, ifPartitionNotExists: Boolean): LogicalPlan = {
     ReflectUtil.createInsertInto(table, partition, Seq.empty[String], query, overwrite, ifPartitionNotExists)
   }
-
-  override def createLike(left: Expression, right: Expression): Expression = {
-    new Like(left, right)
-  }
 }
